Rename files to clean up registry subpackage exports.
No real code changes here; just reorganization and import adjustments.

This adheres to the proposal in RFC-653 (which does not violate our
current conventions - it just tightens them) to avoid exporting
the names of modules whose existence should be considered an
implementation detail.
