+++ python/lsst/daf/butler/registry/__init__.py
@@ -19,16 +19,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-__all__ = [
-    "AmbiguousDatasetError",
-    "DbAuth",
-    "DbAuthError",
-    "ConflictingDefinitionError",
-    "OrphanedRecordError",
-    "Registry",
-]
+from ._config import *
+from ._registry import *
+from ._dbAuth import *
 
-from .config import *
-from .registry import *
-from .dbAuth import *
-from .interfaces import *
+from . import interfaces
