+++ python/lsst/daf/butler/registry/_config.py
