+++ python/lsst/daf/butler/registry/_dbAuth.py
