+++ python/lsst/daf/butler/registry/interfaces/__init__.py
@@ -19,4 +19,4 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .database import *
+from ._database import *
