+++ python/lsst/daf/butler/registry/interfaces/_database.py
