+++ python/lsst/daf/butler/registry/tests/__init__.py
@@ -19,5 +19,5 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .database import *
-from .registry import *
+from ._database import *
+from ._registry import *
